# database-service/openapi.yaml
openapi: 3.0.3
info:
  title: Project Encourage Database Service
  version: 1.0.0
  description: REST API for all database operations in the Project Encourage ecosystem
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Question:
      type: object
      properties:
        id:                { type: integer }
        type:              { type: string, enum: [truth, dare] }
        question:          { type: string }
        user_id:           { type: string }
        server_id:         { type: string }
        is_approved:       { type: boolean }
        approved_by:       { type: string, nullable: true }
        datetime_approved: { type: string, format: date-time, nullable: true }
        is_banned:         { type: boolean }
        ban_reason:        { type: string, nullable: true }
        banned_by:         { type: string, nullable: true }
        datetime_banned:   { type: string, format: date-time, nullable: true }
        is_deleted:        { type: boolean }
        datetime_deleted:  { type: string, format: date-time, nullable: true }
        message_id:        { type: string, nullable: true }
        created:           { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string }
paths:
...
  /question/{id}:
    get:
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The question
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '400':
          description: Invalid ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Question not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    patch:
      summary: Update question fields
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Question' }
      responses:
        '200':
          description: Updated question
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '400':
          description: Invalid ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Question not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...